# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Project Bridge - Truth Ledger (Phase 1)
  description: |
    Universal Truth Ledger API for Retail, Services, and Rentals.
    
    ## The 6 Final Locks
    1. **total_amount is DERIVED** - Backend recalculates from lines, rejects mismatches
    2. **POSTED is immutable** - No updates/deletes once status is POSTED
    3. **Reversals are first-class** - Reversal transactions reference original with negative amounts
    4. **Phone uniqueness is PER TENANT** - UNIQUE (tenant_id, phone_number)
    5. **Metadata is READABLE** - Keys must be snake_case, no nested chaos
    6. **Every table has created_by** - Audit trail for all records
  version: 1.0.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TransactionLine:
      type: object
      required: [description, quantity, unit_price, account_code]
      properties:
        id:
          type: string
          format: uuid
          description: Auto-generated UUID
        description:
          type: string
          description: Human readable name (e.g., "Nike Shoes", "DJ Consultation")
        sku:
          type: string
          description: Stock keeping unit for retail items
        quantity:
          type: number
          description: Item count for retail, hours for services, days for rentals
        unit_price:
          type: number
          description: Price per unit in cents (e.g., 5000 = 50.00)
        total_line_amount:
          type: number
          description: Calculated as quantity * unit_price
        account_code:
          type: string
          description: Chart of accounts mapping (e.g., "200-SALES", "400-RENTAL-INCOME")
        metadata:
          type: object
          description: |
            Type-specific details:
            - RETAIL: {sku, variant, weight_kg, stock_location}
            - SERVICE: {hours, project_code, start_time, end_time, skill_level}
            - RENTAL: {return_date, serial_number, deposit_held, condition_out, insurance_waiver}
    Transaction:
      type: object
      required: [type, currency_code, lines]
      properties:
        id:
          type: string
          format: uuid
          description: Auto-generated UUID (primary key)
        tenant_id:
          type: string
          format: uuid
          description: Strict multi-tenancy isolation
        reference:
          type: string
          description: External reference number (optional)
        entity_id:
          type: string
          format: uuid
          description: Link to entities table (Customer/Supplier)
        created_by_user_id:
          type: string
          format: uuid
          description: User who created this transaction (LOCK 6)
        type:
          type: string
          enum: [RETAIL, SERVICE, RENTAL, EXPENSE]
          description: The Major Switch - determines business logic
        status:
          type: string
          enum: [DRAFT, POSTED, REVERSED, RECONCILED, VOIDED, ARCHIVED]
          description: |
            DRAFT → editable
            POSTED → immutable (LOCK 2)
            RECONCILED → immutable + payment-linked
        payment_status:
          type: string
          enum: [PENDING, PARTIAL, PAID, OVERDUE]
        total_amount:
          type: number
          description: DERIVED from sum of line amounts (LOCK 1)
        currency_code:
          type: string
          description: ISO 4217 (KES, USD, NGN)
        transaction_date:
          type: string
          format: date-time
          description: The economic event date
        reversed_transaction_id:
          type: string
          format: uuid
          description: For reversals - references original transaction (LOCK 3)
        metadata:
          type: object
          description: Additional transaction-level data (snake_case keys only - LOCK 5)
        lines:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLine'
    Entity:
      type: object
      required: [type, display_name]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [CUSTOMER, SUPPLIER, EMPLOYEE]
        display_name:
          type: string
        phone_number:
          type: string
          description: E.164 format (+254...) - UNIQUE per tenant (LOCK 4)
        metadata:
          type: object
          description: Store "Trust Score", "Location", "Notes"
        created_by_user_id:
          type: string
          format: uuid
          description: User who created this entity (LOCK 6)
paths:
  /transactions:
    post:
      summary: Create a transaction (DRAFT)
      description: |
        Creates a new transaction in DRAFT status.
        
        **LOCK 1**: total_amount is calculated from lines, not trusted from input.
        **LOCK 6**: created_by_user_id is required for audit trail.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Validation error
        '403':
          description: Forbidden - attempting to modify immutable transaction
    get:
      summary: List transactions
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, POSTED, REVERSED, RECONCILED, VOIDED, ARCHIVED]
        - in: query
          name: type
          schema:
            type: string
            enum: [RETAIL, SERVICE, RENTAL, EXPENSE]
        - in: query
          name: date_start
          schema:
            type: string
            format: date-time
        - in: query
          name: date_end
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /transactions/{id}:
    get:
      summary: Get transaction detail
      description: Returns full Header + Lines + Entity details
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
  /entities/{id}/transactions:
    get:
      summary: Get all transactions for an entity
      description: History view - all transactions linked to a specific Customer/Supplier
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /transactions/export/csv:
    get:
      summary: Export transactions to CSV
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /entities:
    post:
      summary: Create an entity (Customer/Supplier/Employee)
      description: |
        **LOCK 4**: Phone number uniqueness is enforced per tenant.
        **LOCK 5**: Metadata keys must be snake_case.
        **LOCK 6**: created_by_user_id is required.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entity'
      responses:
        '201':
          description: Created
        '400':
          description: Validation error - phone number already exists for tenant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /transactions/{id}/reverse:
    post:
      summary: Create a reversal transaction for a POSTED transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason_code:
                  type: string
                reason_text:
                  type: string
      responses:
        '201':
          description: Reversal created
        '400':
          description: Invalid request
  /payments:
    post:
      summary: Create a payment
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reference: { type: string }
                amount: { type: number }
                currency_code: { type: string }
      responses:
        '201':
          description: Created
  /payments/{id}/apply:
    post:
      summary: Apply a payment to a transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transaction_id: { type: string }
                amount: { type: number }
      responses:
        '200':
          description: Applied